requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor-5==0.2.18
django-environ==0.12.0
djangorestframework==3.16.0
pillow==11.2.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
